+++ transpiler/src/ast/core/index.ts
@@ -16,6 +16,7 @@ const migratedTypes = [
   'propsDeclaration',
   'valueObjectDeclaration',
   'entityDeclaration',
+  'aggregateDeclaration',
   'domainRuleDeclaration',
 ];
